Initial AltContainer refactoring
