+++ webapp/src/main/resources/public/js/actions/users/UserActions.js
@@ -7,6 +7,9 @@ class UserActions {
             "openNewUserModal",
             "openEditUserModal",
             "openDeleteUserModal",
+            "prevPage",
+            "nextPage",
+            "updatePageSize",
             "reloadCurrentPage",
             "checkUsernameTaken",
             "checkUsernameTakenSuccess",
